requirements.txt
country_list==0.1.3
numpy==1.16.4
colorama==0.4.1
matplotlib==3.1.0
beautifultable==0.8.0
Jinja2==2.10.1
pandas==0.24.2
pandas_profiling==2.3.0
ipython==7.9.0
PyYAML==5.1.2
